+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/ProfilerControllerTest.kt
@@ -6,6 +6,7 @@ import org.hamcrest.Matchers.greaterThan
 import org.json.JSONObject
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mozilla.geckoview.ProfilerController
 import java.io.BufferedReader
 import java.io.ByteArrayInputStream
 import java.io.InputStreamReader
@@ -16,8 +17,8 @@ class ProfilerControllerTest : BaseSessionTest() {
 
     @Test
     fun startAndStopProfiler() {
-        sessionRule.runtime.profilerController.startProfiler(arrayOf<String>(), arrayOf<String>())
-        val result = sessionRule.runtime.profilerController.stopProfiler()
+        ProfilerController.startProfiler(arrayOf<String>(), arrayOf<String>())
+        val result = ProfilerController.stopProfiler()
         val byteArray = sessionRule.waitForResult(result)
         val head = (byteArray[0].toInt() and 0xff) or (byteArray[1].toInt() shl 8 and 0xff00)
         assertThat(